building the regexp capture for CHANGE/MODIFY

+++ parse.py
@@ -43,7 +43,7 @@ ALTER TABLE `test` DROP PRIMARY KEY;
 #statement="""create table test_pk (id int ,PRIMARY KEY  (id) ); """
 #statement="""alter table test change   date_create_new date_create_new timestamp;"""
 statement="""ALTER TABLE `test_table` MODIFY `test_column` bigint(20) DEFAULT NULL; 
-ALTER TABLE table2 CHANGE column1 column2 DOUBLE;"""
+ALTER TABLE table2 CHANGE column1 column2 bigint(20);"""
 
 token_sql=sql_token()
 token_sql.parse_sql(statement)
